Safari-611.1.10/LayoutTests/webgpu/whlsl/logical-negation.html

<!DOCTYPE html>
<html>
<meta charset=utf-8>
<meta name="timeout" content="long">
<title>Lvalues.</title>
<script src="js/test-harness.js"></script>
<script src="../js/webgpu-functions.js"></script>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
const whlslTests = {};

whlslTests.logicalNegation = async () =>
{
    let program = "bool foo(bool x) { return !x; }";
    assert_equals(await callBoolFunction(program, "foo", [makeBool(true)]), false);
    assert_equals(await callBoolFunction(program, "foo", [makeBool(false)]), true);
}

runTests(whlslTests);
</script>
</html>
